Move Animator for sprites to graphics library.
